DDT-466: Rename OVS to JIT
