+++ ui/nginx.d/default.conf.template
@@ -33,11 +33,11 @@ server {
     }
 
     location ~* /api(/.*) {
-        proxy_pass ${ovs_backend}$1;
+        proxy_pass ${jit_backend}$1;
         proxy_set_header Host localhost;
         proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
         proxy_set_header X-Forwarded-Proto "https";
-        proxy_set_header X-Forwarded-Host "${ovs_backend}";
+        proxy_set_header X-Forwarded-Host "${jit_backend}";
         proxy_http_version 1.1;
         proxy_set_header Upgrade $http_upgrade;
         proxy_set_header Connection "Upgrade";
