+++ ui/src/app/controller/services/jit/operations-event.service.ts
@@ -1,9 +1,9 @@
 import {Injectable} from '@angular/core';
 import {HttpClient} from "@angular/common/http";
 import {Observable} from "rxjs";
-import {OperationsEvent} from "../../../model/ovs/operations-event";
+import {OperationsEvent} from "../../../model/jit/operations-event";
 import {Globals} from 'src/app/model/portCall/globals';
-import { TimestampInfo } from 'src/app/model/ovs/timestampInfo';
+import { TimestampInfo } from 'src/app/model/jit/timestampInfo';
 
 @Injectable({
   providedIn: 'root'
@@ -14,8 +14,8 @@ export class OperationsEventService {
   private readonly EVENT_DELIVERY_STATUS_URL: string;
 
   constructor(private httpClient: HttpClient, globals: Globals) {
-    this.TIMESTAMPS_URL = globals.config.ovsBackendURL + "/timestamps";
-    this.OPERATIONS_EVENT_URL = globals.config.ovsBackendURL + "/events";
+    this.TIMESTAMPS_URL = globals.config.jitBackendURL + "/timestamps";
+    this.OPERATIONS_EVENT_URL = globals.config.jitBackendURL + "/events";
     this.EVENT_DELIVERY_STATUS_URL = globals.config.uiSupportBackendURL + "/unofficial/timestamp-info";
   }
