+++ ui/src/app/model/portCall/globals.ts
@@ -1,7 +1,7 @@
 import {Injectable} from '@angular/core';
 import {Port} from "./port";
 import {Terminal} from "./terminal";
-import {Config} from "../ovs/config";
+import {Config} from "../jit/config";
 
 @Injectable()
 export class Globals {
