+++ ui/src/assets/config.json
@@ -26,7 +26,7 @@
   },
   "publisherRoles": ["CA", "AG", "VSL", "TR", "ATH", "PLT", "TWG", "BUK", "LSH"],
   "uiSupportBackendURL": "http://localhost:9091/v1",
-  "ovsBackendURL": "http://localhost:9090/v2",
+  "jitBackendURL": "http://localhost:9090/v1",
   "dateTimeFormat": "dd MMM yyyy HH:mm",
   "cognitoUserPoolId": "eu-west-1_LsLTYEjn5",
   "cognitoAppClientId": "4u49pjh38su1ccea0l4dras9if",
